Simplify the decorator setup substantially - WIP
