+++ promptmage/run.py
@@ -1,0 +1,7 @@
+class MageRun:
+    """This represents a run of a flow in the PromptMage."""
+
+    def __init__(
+        self,
+    ):
+        pass
